Add new emergency skin location
This change enables the new SkinSelector code to find and use the emergency skin.  The new skin selector code treats all skins equally such that every skin now has a skin.xml at the top level of the skin directory.

All changes to the emergency skin located at the top level of this repository should also be made to this file.